fix: rename TextFieldState enum
