+++ src/components/CircleIconBtn/index.ts
@@ -1,6 +1,6 @@
 import { MetCircleIconBtn } from "./CircleIconBtn";
 import { MetCircleIconBtnProps } from "./CircleIconBtnProps";
-import { SizeCircleIconBtn } from "./sizeCircleIconBtn";
+import { SizeCircleIconBtn } from "./sizeCircleIconBtn.enum";
 
 export { MetCircleIconBtnProps, SizeCircleIconBtn };
