+++ src/components/PromptInfo/PromptInfo.tsx
@@ -4,7 +4,7 @@ import React, { FC, ReactElement, useEffect, useRef } from "react";
 import styled from "styled-components";
 
 import { colors } from "../styles/colors.js";
-import { SizeCircleIconBtn as Size } from "../CircleIconBtn/sizeCircleIconBtn";
+import { SizeCircleIconBtn as Size } from "../CircleIconBtn/sizeCircleIconBtn.enum";
 
 import { MetPromptInfoProps } from "./index";
