+++ src/components/PromptInfo/index.ts
@@ -1,6 +1,6 @@
 import { MetPromptInfo } from "./PromptInfo";
 import { MetPromptInfoProps } from "./PromptInfoProps";
-import { SizeCircleIconBtn } from "../CircleIconBtn/sizeCircleIconBtn";
+import { SizeCircleIconBtn } from "../CircleIconBtn/sizeCircleIconBtn.enum";
 
 export { MetPromptInfoProps, SizeCircleIconBtn };
