+++ src/components/TextField/Input/Input.stories.tsx
@@ -4,7 +4,7 @@ import { Meta } from "@storybook/react/types-6-0";
 
 import MetTextField, {
     MetInputProps,
-    textFieldState as stateTF,
+    TextFieldState as stateTF,
 } from "./index";
 
 import {
@@ -49,11 +49,11 @@ export const Default = () => {
         state: select(
             "State",
             {
-                Default: stateTF.default,
-                Error: stateTF.error,
-                Success: stateTF.success,
+                Default: stateTF.DEFAULT,
+                Error: stateTF.ERROR,
+                Success: stateTF.SUCCESS,
             },
-            stateTF.default
+            stateTF.DEFAULT
         ),
         hoverColor: color("Hover color", undefined),
         focusColor: color("Focus color", undefined),
