--- src/components/TextField/TextFieldState.ts
@@ -1,5 +1,0 @@
-export enum textFieldState {
-  error = "error",
-  success = "success",
-  default = "default",
-}
