+++ src/components/TextField/textFieldState.enum.ts
@@ -1,0 +1,5 @@
+export enum TextFieldState {
+  ERROR,
+  SUCCESS,
+  DEFAULT,
+}
